refactor: replace camelCase with kebab-case in file names
