+++ src/app/login/page.tsx
@@ -1,5 +1,5 @@
-import Header from "@/components/header/header";
-import LoginForm from "@/components/loginForm/loginForm";
+import Header from "@/components/login-header/login-header";
+import LoginForm from "@/components/login-form/login-form";
 
 export default function LoginPage() {
   return (
