+++ src/components/login-header/login-header.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import logoImg from "../../../public/icon.svg";
-import styles from "./header.module.scss";
+import styles from "./login-header.module.scss";
 
 export default function Header() {
   return (
